+++ game-silenthill2/index.js
@@ -3,29 +3,39 @@ Name: Silent Hill 2 Remake Vortex Extension
 Structure: UE5
 Author: ChemBoy1
 Version: 0.3.0
-Date: 2026-02-01
+Date: 2026-02-02
 //////////////////////////////////////////////////*/
 
 //Import libraries
 const { actions, fs, util, selectors, log } = require('vortex-api');
 const path = require('path');
 const template = require('string-template');
+const { parseStringPromise } = require('xml2js');
 
 //Specify all information about the game
 const GAME_ID = "silenthill2";
 const STEAMAPP_ID = "2124490";
 const EPICAPP_ID = "c4dc308a1b69492aba4d47f7feaa1083";
 const GOGAPP_ID = "2051029707";
-const XBOXAPP_ID = null;
-const XBOXEXECNAME = null;
+const XBOXAPP_ID = ""; //There is an Xbox version, but not on Game Pass, so cannot get data
+const XBOXEXECNAME = "";
 const EPIC_CODE_NAME = "SHProto";
 const EXEC_FOLDER_NAME = "Win64";
 //const EXEC_FOLDER_NAME = "WinGDK"; //XBOX Version
 const GAME_NAME = "Silent Hill 2 (2024)";
 const GAME_NAME_SHORT = "Silent Hill 2";
 const EXEC = "SHProto.exe";
+const EXEC_XBOX = `gamelaunchhelper.exe`;
+const APPMANIFEST_FILE = 'appxmanifest.xml';
 //const EXEC = XBOXEXECNAME; //XBOX Version
 const IO_STORE = true; //true if the Paks folder contains .ucas and .utoc files
+const PCGAMINGWIKI_URL = "https://www.pcgamingwiki.com/wiki/Silent_Hill_2";
+const EXTENSION_URL = "https://www.nexusmods.com/site/mods/963"; //Nexus link to this extension. Used for links
+
+let GAME_PATH = '';
+let STAGING_FOLDER = '';
+let DOWNLOAD_FOLDER = '';
+let GAME_VERSION = '';
 
 /*
   Unreal Engine Game Data
@@ -38,15 +48,16 @@ const UNREALDATA = {
   fileExt: ['.pak', '.ucas', '.utoc'],
   loadOrder: true,
 }
+const UE4SS_MOD_PATH = path.join('ue4ss', 'Mods');
 
 //This information will be filled in from the data above
-const SHIPPING_EXE = `${EPIC_CODE_NAME}\\Binaries\\${EXEC_FOLDER_NAME}\\${EPIC_CODE_NAME}-${EXEC_FOLDER_NAME}-Shipping.exe`;
+const SHIPPING_EXE = path.join(EPIC_CODE_NAME, 'Binaries', EXEC_FOLDER_NAME, `${EPIC_CODE_NAME}-${EXEC_FOLDER_NAME}-Shipping.exe`);
 const BINARIES_ID = `${GAME_ID}-binaries`;
-const EXEC_PATH = `${EPIC_CODE_NAME}\\Binaries\\${EXEC_FOLDER_NAME}`;
+const BINARIES_PATH = path.join(EPIC_CODE_NAME, 'Binaries', EXEC_FOLDER_NAME);
 
 const ROOT_ID = `${GAME_ID}-root`;
 const ROOT_FILE = EPIC_CODE_NAME;
-const ROOT_IDX = `${EPIC_CODE_NAME}\\`;
+const ROOT_IDX = `${EPIC_CODE_NAME}${path.sep}`;
 
 const UE5_ID = `${GAME_ID}-ue5`;
 const UE5_ALT_ID = `${GAME_ID}-pakalt`;
@@ -91,13 +102,28 @@ if (USERID_FOLDER === undefined) {
 const SAVE_PATH = path.join(SAVE_FOLDER, USERID_FOLDER);
 
 const SCRIPTS_ID = `${GAME_ID}-scripts`;
-const SCRIPTS_PATH = `${EPIC_CODE_NAME}\\Binaries\\${EXEC_FOLDER_NAME}\\ue4ss\\Mods`;
+const SCRIPTS_PATH = path.join(BINARIES_PATH, UE4SS_MOD_PATH);
 const SCRIPTS_EXT = ".lua";
 const SCRIPTS_FILE = "Scripts";
-const SCRIPTS_IDX = `Scripts\\`;
+const SCRIPTS_IDX = `Scripts${path.sep}`;
+
+const DLL_ID = `${GAME_ID}-ue4ssdll`;
+const DLL_NAME = "UE4SS DLL Mod";
+const DLL_EXT = ".dll";
+const DLL_FOLDER = "dlls";
+const DLL_PATH = SCRIPTS_PATH;
 
 const UE4SS_ID = `${GAME_ID}-ue4ss`;
+const UE4SS_NAME = "UE4SS";
 const UE4SS_FILE = "dwmapi.dll";
+const UE4SS_DLFILE_STRING = "ue4ss";
+const UE4SS_URL = "https://github.com/UE4SS-RE/RE-UE4SS/releases";
+const UE4SS_SETTINGS_FILE = 'UE4SS-settings.ini';
+const UE4SS_SETTINGS_FILEPATH = path.join('ue4ss', UE4SS_SETTINGS_FILE); //relative to Binaries folder
+const UE4SS_MODSJSON_FILE = 'mods.json';
+const UE4SS_MODSTXT_FILE = 'mods.txt';
+const UE4SS_MODSJSON_FILEPATH = path.join(UE4SS_MOD_PATH, UE4SS_MODSJSON_FILE); //relative to Binaries folder
+const UE4SS_MODSTXT_FILEPATH = path.join(UE4SS_MOD_PATH, UE4SS_MODSTXT_FILE);
 
 //Set file number for pak installer file selection (needs to be 3 if IO Store is used to accomodate .ucas and .utoc files)
 let PAK_FILE_MIN = 1;
@@ -119,7 +145,7 @@ const spec = {
     "executable": EXEC,
     "logo": `${GAME_ID}.jpg`,
     "mergeMods": true,
-    "modPath": EXEC_PATH,
+    "modPath": BINARIES_PATH,
     "modPathIsRelative": true,
     "requiredFiles": [
       EXEC,
@@ -143,13 +169,19 @@ const spec = {
       "id": SCRIPTS_ID,
       "name": "UE4SS Scripts",
       "priority": "high",
-      "targetPath": `{gamePath}\\${SCRIPTS_PATH}`
+      "targetPath": path.join('{gamePath}', SCRIPTS_PATH)
+    },
+    { //UE4SS DLL Mods
+      "id": DLL_ID,
+      "name": DLL_NAME,
+      "priority": "high",
+      "targetPath": path.join('{gamePath}', DLL_PATH)
     },
     {
       "id": LOGICMODS_ID,
       "name": "UE4SS LogicMods (Blueprint)",
       "priority": "high",
-      "targetPath": `{gamePath}\\${LOGICMODS_PATH}`
+      "targetPath": path.join('{gamePath}', LOGICMODS_PATH)
     },
     {
       "id": UE4SSCOMBO_ID,
@@ -168,7 +200,7 @@ const spec = {
       "id": SAVE_ID,
       "name": "Saves (LocalAppData)",
       "priority": "high",
-      "targetPath": `${SAVE_PATH}\\${USERID_FOLDER}`
+      "targetPath": path.join(SAVE_PATH, USERID_FOLDER)
     },
     */
     {
@@ -181,32 +213,32 @@ const spec = {
       "id": UE5_ID,
       "name": "UE5 Paks",
       "priority": "high",
-      "targetPath": `{gamePath}\\${UE5_PATH}`
+      "targetPath": path.join('{gamePath}', UE5_PATH)
     },
     {
       "id": UE5_ALT_ID,
       "name": 'UE5 Paks (no "~mods")',
       "priority": "high",
-      "targetPath": `{gamePath}\\${UE5_ALT_PATH}`
+      "targetPath": path.join('{gamePath}', UE5_ALT_PATH)
     },
     {
       "id": BINARIES_ID,
       "name": "Binaries (Engine Injector)",
       "priority": "high",
-      "targetPath": `{gamePath}\\${EXEC_PATH}`
+      "targetPath": path.join('{gamePath}', BINARIES_PATH)
     },
     {
       "id": UE4SS_ID,
       "name": "UE4SS",
       "priority": "low",
-      "targetPath": `{gamePath}\\${EXEC_PATH}`
+      "targetPath": path.join('{gamePath}', BINARIES_PATH)
     },
   ],
   "discovery": {
     "ids": [
       STEAMAPP_ID,
       EPICAPP_ID,
-      //GOGAPP_ID,
+      GOGAPP_ID,
       //XBOXAPP_ID
     ],
     "names": []
@@ -215,22 +247,40 @@ const spec = {
 
 //3rd party tools and launchers
 const tools = [
-  /*
   {
-    id: "LaunchModdedGame",
-    name: "Launch Modded Game",
+    id: `${GAME_ID}-customlaunch`,
+    name: "Custom Launch",
     logo: `exec.png`,
     executable: () => EXEC,
     requiredFiles: [EXEC],
     detach: true,
     relative: true,
     exclusive: true,
-    defaultPrimary: true,
-    isPrimary: true,
-    parameters: ['-fileopenlog']
-  },
-  */
+    //defaultPrimary: true,
+    //parameters: ['-fileopenlog']
+  }, //*/
 ];
+
+// BASIC EXTENSION FUNCTIONS //////////////////////////////////////////////////////
+
+function statCheckSync(gamePath, file) {
+  try {
+    fs.statSync(path.join(gamePath, file));
+    return true;
+  }
+  catch (err) {
+    return false;
+  }
+}
+async function statCheckAsync(gamePath, file) {
+  try {
+    await fs.statAsync(path.join(gamePath, file));
+    return true;
+  }
+  catch (err) {
+    return false;
+  }
+}
 
 //Set mod type priorities
 function modTypePriority(priority) {
@@ -258,66 +308,82 @@ function makeGetModPath(api, gameSpec) {
     : pathPattern(api, gameSpec.game, gameSpec.game.modPath);
 }
 
-//Find game information by API utility
-async function queryGame() {
-  let game = await util.GameStoreHelper.findByAppId(spec.discovery.ids);
-  return game;
+//Find game installation directory
+function makeFindGame(api, gameSpec) {
+  return () => util.GameStoreHelper.findByAppId(gameSpec.discovery.ids)
+    .then((game) => game.gamePath);
 }
 
-//Find game install location 
-async function queryPath(api) {
-  let game = await queryGame();
-  return game.gamePath;
-}
-
-//Set launcher requirements
-async function requiresLauncher() {
-//async function requiresLauncher(api) {
-  let game = await queryGame();
-
-  if (game.gameStoreId === "steam") {
-    return undefined;
-  }
-
-  if (game.gameStoreId === "gog") {
-    return undefined;
-  }
-  if (game.gameStoreId === "epic") {
-    //api.showErrorNotification('Epic AppName needed to launch game', 'Please see the extension page on Nexus Mods for instructions and send the Epic AppName to the mod author.');
-    //return undefined;
-    return {
-      launcher: "epic",
-      addInfo: {
-        appId: EPICAPP_ID,
-      },
-    };
+async function requiresLauncher(gamePath, store) {
+  /*if (store === 'xbox') {
+      return Promise.resolve({
+          launcher: 'xbox',
+          addInfo: {
+              appId: XBOXAPP_ID,
+              parameters: [{ appExecName: XBOXEXECNAME }],
+          },
+      });
   } //*/
-
-  if (game.gameStoreId === "xbox") {
-    return {
-      launcher: "xbox",
-      addInfo: {
-        appId: XBOXAPP_ID,
-        // appExecName is the <Application id="" in the appxmanifest.xml file
-        parameters: [{ appExecName: XBOXEXECNAME }],
-      },
-    };
-  }
-
-  return undefined;
+  if (store === 'epic') {
+    return Promise.resolve({
+        launcher: 'epic',
+        addInfo: {
+            appId: EPICAPP_ID,
+        },
+    });
+  } //*/
+  return Promise.resolve(undefined);
 }
 
-//Test for save files
+//Get correct executable for game version
+function getExecutable(discoveryPath) {
+  if (statCheckSync(discoveryPath, EXEC_XBOX)) {
+    GAME_VERSION = 'xbox';
+    return EXEC_XBOX;
+  };
+  //add GOG/EGS/Demo versions here if needed
+  GAME_VERSION = 'default';
+  return EXEC;
+}
+
+//Get correct game version
+async function setGameVersionPath(discoveryPath) {
+  if (statCheckAsync(discoveryPath, EXEC_XBOX)) {
+    GAME_VERSION = 'xbox';
+    return GAME_VERSION;
+  };
+  //add GOG/EGS/Demo versions here if needed
+  GAME_VERSION = 'default';
+  return GAME_VERSION;
+}
+
+const getDiscoveryPath = (api) => { //get the game's discovered path
+  const state = api.getState();
+  const discovery = util.getSafe(state, [`settings`, `gameMode`, `discovered`, GAME_ID], {});
+  return discovery === null || discovery === void 0 ? void 0 : discovery.path;
+};
+
+async function purge(api) {
+  return new Promise((resolve, reject) => api.events.emit('purge-mods', true, (err) => err ? reject(err) : resolve()));
+}
+async function deploy(api) {
+  return new Promise((resolve, reject) => api.events.emit('deploy-mods', (err) => err ? reject(err) : resolve()));
+}
+
+// MOD INSTALLER FUNCTIONS ///////////////////////////////////////////////////
+
+//Test for UE4SS combo (pak and lua/dll) mod files
 function testUe4ssCombo(files, gameId) {
-  const isMod = files.find(file => path.extname(file).toLowerCase() === SCRIPTS_EXT) !== undefined;
-  const isMod2 = files.find(file => path.extname(file).toLowerCase() === LOGICMODS_EXT) !== undefined;
-  const isFolder = files.find(file => path.basename(file) === ROOT_FILE) !== undefined;
-  let supported = (gameId === spec.game.id) && isMod && isMod2 && isFolder;
+  const isMod = files.some(file => (path.extname(file).toLowerCase() === SCRIPTS_EXT));
+  const isModAlt = files.some(file => (path.basename(file).toLowerCase() === 'binaries')); //added to catch mods packaged with paks and dll/asi, but no lua scripts.
+  const isMod2 = files.some(file => (path.extname(file).toLowerCase() === LOGICMODS_EXT));
+  const isFolder = files.some(file => (path.basename(file) === ROOT_FILE));
+  let supported = (gameId === spec.game.id) && ( isMod || isModAlt ) && isMod2 && isFolder;
 
   // Test for a mod installer
   if (supported && files.find(file =>
-      (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
-      (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
+    (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
+    (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
     supported = false;
   }
 
@@ -327,7 +393,7 @@ function testUe4ssCombo(files, gameId) {
   });
 }
 
-//Install save files
+//Install combo files
 function installUe4ssCombo(files, fileName) {
   const modFile = files.find(file => path.basename(file) === ROOT_FILE);
   const idx = modFile.indexOf(ROOT_IDX);
@@ -472,6 +538,71 @@ function installScripts(files, fileName) {
   return Promise.resolve({ instructions });
 }
 
+
+//Test for UE4SS DLL files
+function testDll(files, gameId) {
+  const isMod = files.some(file => (path.extname(file).toLowerCase() === DLL_EXT));
+  const isFolder = files.some(file => (path.basename(file).toLowerCase() === DLL_FOLDER.toLowerCase()));
+  let supported = (gameId === spec.game.id) && isMod && isFolder;
+
+  // Test for a mod installer
+  if (supported && files.find(file =>
+      (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
+      (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
+    supported = false;
+  }
+
+  return Promise.resolve({
+    supported,
+    requiredFiles: [],
+  });
+}
+
+//Install UE4SS DLL files
+function installDll(files, fileName) {
+  const modFile = files.find(file => (path.basename(file).toLowerCase() === DLL_FOLDER.toLowerCase()));
+  const idx = modFile.indexOf(`${path.basename(modFile)}${path.sep}`);
+  const rootPath = path.dirname(modFile);
+  const setModTypeInstruction = { type: 'setmodtype', value: DLL_ID };
+  const MOD_NAME = path.basename(fileName);
+  let MOD_FOLDER = path.basename(rootPath);
+  if (MOD_FOLDER === '.') {
+    MOD_FOLDER = MOD_NAME.replace(/(\.installing)*(\.zip)*(\.rar)*(\.7z)*( )*/gi, '');
+  }
+  
+  const ENABLEDTXT_FILE = 'enabled.txt'
+  const ENABLEDTXT_PATH = path.join(fileName, rootPath, ENABLEDTXT_FILE);
+  try {
+    fs.statSync(ENABLEDTXT_PATH);
+  } catch (err) {
+    try {
+      fs.writeFileSync(
+        ENABLEDTXT_PATH,
+        ``,
+        { encoding: "utf8" },
+      );
+      files.push(path.join(rootPath, ENABLEDTXT_FILE));
+      log('info', `Successfully created enabled.txt for UE4SS DLL Mod: ${MOD_NAME}`);
+    } catch (err) {
+      log('error', `Could not create enabled.txt for UE4SS DLL Mod: ${MOD_NAME}`);
+    }
+  }
+
+  //Filter files and set instructions
+  const filtered = files.filter(file =>
+    ((file.indexOf(rootPath) !== -1) && (!file.endsWith(path.sep)))
+  );
+  const instructions = filtered.map(file => {
+    return {
+      type: 'copy',
+      source: file,
+      destination: path.join(MOD_FOLDER, file.substr(idx)),
+    };
+  });
+  instructions.push(setModTypeInstruction);
+  return Promise.resolve({ instructions });
+}
+
 //Test for config files
 function testConfig(files, gameId) {
   const isConfig = files.find(file => CONFIG_FILES.includes(path.basename(file).toLowerCase())) !== undefined;
@@ -593,6 +724,8 @@ function installSave(files) {
   return Promise.resolve({ instructions });
 }
 
+// AUTO-DOWNLOADER FUNCTIONS ///////////////////////////////////////////////////
+
 //Check if UE4SS is installed
 function isUe4ssInstalled(api, spec) {
   const state = api.getState();
@@ -600,130 +733,79 @@ function isUe4ssInstalled(api, spec) {
   return Object.keys(mods).some(id => mods[id]?.type === UE4SS_ID);
 }
 
-//Function to auto-download UE4SS from Nexus
+//* Download UE4SS from GitHub page (user browse for download)
 async function downloadUe4ss(api, gameSpec) {
-  let modLoaderInstalled = isUe4ssInstalled(api, gameSpec);
-  
-  if (!modLoaderInstalled) {
-    //notification indicating install process
-    const MOD_NAME = `UE4SS`;
-    const NOTIF_ID = `${GAME_ID}-${MOD_NAME}-installing`;
-    api.sendNotification({
-      id: NOTIF_ID,
-      message: `Installing ${MOD_NAME}`,
-      type: 'activity',
-      noDismiss: true,
-      allowSuppress: false,
-    });
-    //make sure user is logged into Nexus Mods account in Vortex
-    if (api.ext?.ensureLoggedIn !== undefined) {
-      await api.ext.ensureLoggedIn();
-    }
+  let isInstalled = isUe4ssInstalled(api, gameSpec);
+  const URL = UE4SS_URL;
+  const MOD_NAME = UE4SS_NAME;
+  const MOD_TYPE = UE4SS_ID;
+  const ARCHIVE_NAME = UE4SS_DLFILE_STRING;
+  const instructions = api.translate(`Click on Continue below to open the browser. - `
+    + `Navigate to the latest experimental version of ${MOD_NAME} on the GitHub releases page and `
+    + `click on the appropriate file to download and install the mod.`
+  );
 
-    const modPageId = 19;
-    try {
-      //get the mod files information from Nexus
-      const modFiles = await api.ext.nexusGetModFiles(gameSpec.game.id, modPageId);
-      const fileTime = (input) => Number.parseInt(input.uploaded_time, 10);
-      const file = modFiles
-        .filter(file => file.category_id === 1)
-        .sort((lhs, rhs) => fileTime(lhs) - fileTime(rhs))[0];
-      if (file === undefined) {
-        throw new util.ProcessCanceled(`No ${MOD_NAME} main file found`);
-      }
-      //Download the mod
-      const dlInfo = {
-        game: gameSpec.game.id,
-        name: MOD_NAME,
-      };
-      const nxmUrl = `nxm://${gameSpec.game.id}/mods/${modPageId}/files/${file.file_id}`;
-      const dlId = await util.toPromise(cb =>
-        api.events.emit('start-download', [nxmUrl], dlInfo, undefined, cb, undefined, { allowInstall: false }));
-      const modId = await util.toPromise(cb =>
-        api.events.emit('start-install-download', dlId, { allowAutoEnable: false }, cb));
-      const profileId = selectors.lastActiveProfileForGame(api.getState(), gameSpec.game.id);
-      const batched = [
-        actions.setModsEnabled(api, profileId, [modId], true, {
-          allowAutoDeploy: true,
-          installed: true,
-        }),
-        actions.setModType(gameSpec.game.id, modId, UE4SS_ID), // Set the mod type
-      ];
-      util.batchDispatch(api.store, batched); // Will dispatch both actions.
-    //Show the user the download page if the download, install process fails
-    } catch (err) {
-      const errPage = `https://www.nexusmods.com/${gameSpec.game.id}/mods/${modPageId}/files/?tab=files`;
-      api.showErrorNotification(`Failed to download/install ${MOD_NAME}`, err);
-      util.opn(errPage).catch(() => null);
-    } finally {
-      api.dismissNotification(NOTIF_ID);
-    }
-  }
-}
-
-//Notification if Config, Save, and Creations folders are not on the same partition
-function legacyModsNotify(api, legacyMods) {
-  const NOTIF_ID = `${GAME_ID}-legacymodsnotify`;
-  const MESSAGE = 'Reinstall Pak Mods to Make Sortable';
-  api.sendNotification({
-    id: NOTIF_ID,
-    type: 'warning',
-    message: MESSAGE,
-    allowSuppress: true,
-    actions: [
-      {
-        title: 'More',
-        action: (dismiss) => {
-          api.showDialog('question', MESSAGE, {
-            text: `\n\n`
-                + `Due to a bug in a handful of Unreal Engine Vortex game extensions, your pak mods were assigned a modType ID that was shared among several games.\n`
-                + `This bug can result in the Load Order tab not properly load ordering your pak mods.\n`
-                + `A list of the affected mods is shown below. You must Reinstall these mods to make them sortable.\n`
-                + `If you don't Reinstall thes mods, they will still function, but they will sit at the bottom of the loading order and will not be sortable.\n`
-                + `\n`
-                + `Perform the following steps to Reinstall the affected mods:\n
-                  1. Filter your Mods page by Mod Type "Legacy UE - REINSTALL TO SORT" using the categories at the top.\n
-                  2. Use the "CTRL + A" keyboard shortcut to select all displayed mods.\n
-                  3. Click the "Reinstall" button in the blue ribbon at the bottom of the Mods page.\n
-                  4. You can now sort all of your pak mods in the Load Order tab.\n`
-                + `\n`
-                + `Pak Mods to Reinstall:\n` 
-                + `${legacyMods.join('\n')}`
-                + `\n`
-                + `\n`
-          }, [
-            { label: 'Acknowledge', action: () => dismiss() },
-            {
-              label: 'Never Show Again', action: () => {
-                api.suppressNotification(NOTIF_ID);
-                dismiss();
+  if (!isInstalled) {
+    return new Promise((resolve, reject) => { //Browse and download the mod
+      return api.emitAndAwait('browse-for-download', URL, instructions)
+      .then((result) => { //result is an array with the URL to the downloaded file as the only element
+        if (!result || !result.length) { //user clicks outside the window without downloading
+          return reject(new util.UserCanceled());
+        }
+        if (!result[0].toLowerCase().includes(ARCHIVE_NAME)) { //if user downloads the wrong file
+          return reject(new util.ProcessCanceled('Selected wrong download'));
+        } //*/
+        return Promise.resolve(result);
+      })
+      .catch((error) => {
+        return reject(error);
+      })
+      .then((result) => {
+        const dlInfo = {game: gameSpec.game.id, name: MOD_NAME};
+        api.events.emit('start-download', result, {}, undefined,
+          async (error, id) => { //callback function to check for errors and pass id to and call 'start-install-download' event
+            if (error !== null && (error.name !== 'AlreadyDownloaded')) {
+              return reject(error);
+            }
+            api.events.emit('start-install-download', id, { allowAutoEnable: true }, async (error) => { //callback function to complete the installation
+              if (error !== null) {
+                return reject(error);
               }
-            },
-          ]);
-        },
-      },
-    ],
-  });
-}
-
-//Setup function
-async function setup(discovery, api, gameSpec) {
-  //SYNCHRONOUS CODE //////////////////////////////////////
-  const state = api.getState();
-  const mods = util.getSafe(state, ['persistent', 'mods', gameSpec.game.id], {});
-  const legacyMods = Object.keys(mods).filter(id => mods[id]?.type === LEGACY_UE5_SORTABLE_ID);
-  if (legacyMods.length > 0) {
-    legacyModsNotify(api, legacyMods);
+              const profileId = selectors.lastActiveProfileForGame(api.getState(), GAME_ID);
+              const batched = [
+                actions.setModsEnabled(api, profileId, result, true, {
+                  allowAutoDeploy: true,
+                  installed: true,
+                }),
+                actions.setModType(GAME_ID, result[0], MOD_TYPE), // Set the mod type
+              ];
+              util.batchDispatch(api.store, batched); // Will dispatch both actions.
+              return resolve();
+            });
+          }, 
+          'never',
+          { allowInstall: false },
+        );
+      });
+    })
+    .catch(err => {
+      if (err instanceof util.UserCanceled) {
+        api.showErrorNotification(`User cancelled download/install of ${MOD_NAME}. Please re-launch Vortex and try again.`, err, { allowReport: false });
+        //util.opn(URL).catch(() => null);
+        return Promise.resolve();
+      } else if (err instanceof util.ProcessCanceled) {
+        api.showErrorNotification(`Failed to download/install ${MOD_NAME}. Please re-launch Vortex and try again or download manually from modDB at the opened paged and install the zip in Vortex.`, err, { allowReport: false });
+        util.opn(URL).catch(() => null);
+        return Promise.reject(err);
+      } else {
+        return Promise.reject(err);
+      }
+    });
   }
-  //await downloadUe4ss(api, gameSpec);
-  //await fs.ensureDirWritableAsync(path.join(CONFIG_PATH));
-  //await fs.ensureDirWritableAsync(path.join(SAVE_PATH));
-  await fs.ensureDirWritableAsync(path.join(discovery.path, SCRIPTS_PATH));
-  await fs.ensureDirWritableAsync(path.join(discovery.path, LOGICMODS_PATH));
-  return fs.ensureDirWritableAsync(path.join(discovery.path, UE5_PATH));
-}
+} //*/
 
-///*
+// UNREAL FUNCTIONS ///////////////////////////////////////////////////////////////
+
 //UNREAL - Pre-sort function
 async function preSort(api, items, direction) {
   const mods = util.getSafe(api.store.getState(), ['persistent', 'mods', spec.game.id], {});
@@ -873,11 +955,6 @@ function UNREALEXTENSION(context) {
 
 
   const getUnrealModsPath = (game) => {
-    /*
-    const absModsPath = UNREALDATA.absModsPath;
-    if (absModsPath)
-      return absModsPath;
-    */
     const modsPath = UNREALDATA.modsPath;
     const state = context.api.getState();
     const discoveryPath = util.getSafe(state.settings, ['gameMode', 'discovered', game.id, 'path'], undefined);
@@ -901,18 +978,115 @@ function UNREALEXTENSION(context) {
   );
 }
 
+// MAIN FUNCTIONS ///////////////////////////////////////////////////////////////
+
+async function resolveGameVersion(gamePath, exePath) {
+  GAME_VERSION = await setGameVersionPath(gamePath);
+  //SHIPPING_EXE = getShippingExe(gamePath);
+  const READ_FILE = path.join(gamePath, EXEC);
+  let version = '0.0.0';
+  if (GAME_VERSION === 'xbox') { // use appxmanifest.xml for Xbox version
+    try { //try to parse appxmanifest.xml
+      const appManifest = await fs.readFileAsync(path.join(gamePath, APPMANIFEST_FILE), 'utf8');
+      const parsed = await parseStringPromise(appManifest);
+      version = parsed?.Package?.Identity?.[0]?.$?.Version;
+      return Promise.resolve(version);
+    } catch (err) {
+      log('error', `Could not read appmanifest.xml file to get Xbox game version: ${err}`);
+      return Promise.resolve(version);
+    }
+  }
+  else { //use shipping exe (note that this only returns the UE engine version right now)
+    try {
+      const exeVersion = require('exe-version');
+      version = await exeVersion.getProductVersion(READ_FILE);
+      //log('warn', `Resolved game version for ${GAME_ID} to: ${version}`);
+      return Promise.resolve(version); 
+    } catch (err) {
+      log('error', `Could not read ${READ_FILE} file to get Steam game version: ${err}`);
+      return Promise.resolve(version);
+    }
+  }
+}
+
+//Notification if Config, Save, and Creations folders are not on the same partition
+function legacyModsNotify(api, legacyMods) {
+  const NOTIF_ID = `${GAME_ID}-legacymodsnotify`;
+  const MESSAGE = 'Reinstall Pak Mods to Make Sortable';
+  api.sendNotification({
+    id: NOTIF_ID,
+    type: 'warning',
+    message: MESSAGE,
+    allowSuppress: true,
+    actions: [
+      {
+        title: 'More',
+        action: (dismiss) => {
+          api.showDialog('question', MESSAGE, {
+            text: `\n\n`
+                + `Due to a bug in a handful of Unreal Engine Vortex game extensions, your pak mods were assigned a modType ID that was shared among several games.\n`
+                + `This bug can result in the Load Order tab not properly load ordering your pak mods.\n`
+                + `A list of the affected mods is shown below. You must Reinstall these mods to make them sortable.\n`
+                + `If you don't Reinstall thes mods, they will still function, but they will sit at the bottom of the loading order and will not be sortable.\n`
+                + `\n`
+                + `Perform the following steps to Reinstall the affected mods:\n
+                  1. Filter your Mods page by Mod Type "Legacy UE - REINSTALL TO SORT" using the categories at the top.\n
+                  2. Use the "CTRL + A" keyboard shortcut to select all displayed mods.\n
+                  3. Click the "Reinstall" button in the blue ribbon at the bottom of the Mods page.\n
+                  4. You can now sort all of your pak mods in the Load Order tab.\n`
+                + `\n`
+                + `Pak Mods to Reinstall:\n` 
+                + `${legacyMods.join('\n')}`
+                + `\n`
+                + `\n`
+          }, [
+            { label: 'Acknowledge', action: () => dismiss() },
+            {
+              label: 'Never Show Again', action: () => {
+                api.suppressNotification(NOTIF_ID);
+                dismiss();
+              }
+            },
+          ]);
+        },
+      },
+    ],
+  });
+}
+
+//Setup function
+async function setup(discovery, api, gameSpec) {
+  //SYNCHRONOUS CODE //////////////////////////////////////
+  const state = api.getState();
+  GAME_PATH = discovery.path;
+  STAGING_FOLDER = selectors.installPathForGame(state, GAME_ID);
+  DOWNLOAD_FOLDER = selectors.downloadPathForGame(state, GAME_ID);
+  const mods = util.getSafe(state, ['persistent', 'mods', gameSpec.game.id], {});
+  const legacyMods = Object.keys(mods).filter(id => mods[id]?.type === LEGACY_UE5_SORTABLE_ID);
+  if (legacyMods.length > 0) {
+    legacyModsNotify(api, legacyMods);
+  }
+  GAME_VERSION = await setGameVersionPath(GAME_PATH);
+  //await downloadUe4ss(api, gameSpec);
+  //await fs.ensureDirWritableAsync(path.join(CONFIG_PATH));
+  //await fs.ensureDirWritableAsync(path.join(SAVE_PATH));
+  await fs.ensureDirWritableAsync(path.join(GAME_PATH, SCRIPTS_PATH));
+  await fs.ensureDirWritableAsync(path.join(GAME_PATH, LOGICMODS_PATH));
+  return fs.ensureDirWritableAsync(path.join(GAME_PATH, UE5_PATH));
+}
+
 //Let vortex know about the game
 function applyGame(context, gameSpec) {
   //register the game
   const game = {
     ...gameSpec.game,
-    queryPath,
+    queryPath: makeFindGame(context.api, gameSpec),
+    executable: getExecutable,
     queryModPath: makeGetModPath(context.api, gameSpec),
-    requiresLauncher,
-    //requiresLauncher: async () => await requiresLauncher(context.api),
+    requiresLauncher: requiresLauncher,
     requiresCleanup: true,
     setup: async (discovery) => await setup(discovery, context.api, gameSpec),
-    executable: () => gameSpec.game.executable,
+    getGameVersion: resolveGameVersion,
     supportedTools: tools,
   };
   context.registerGame(game);
@@ -932,9 +1106,109 @@ function applyGame(context, gameSpec) {
   //35 is pak installer above
   context.registerInstaller(`${GAME_ID}-ue4ss`, 40, testUe4ss, installUe4ss);
   context.registerInstaller(`${GAME_ID}-ue4ss-scripts`, 45, testScripts, installScripts);
-  context.registerInstaller(`${GAME_ID}-root`, 50, testRoot, installRoot);
+  context.registerInstaller(DLL_ID, 47, testDll, installDll);
+  context.registerInstaller(`${GAME_ID}-root`, 49, testRoot, installRoot);
   //context.registerInstaller(`${GAME_ID}-config`, 55, testConfig, installConfig);
   //context.registerInstaller(`${GAME_ID}-save`, 60, testSave, installSave);
+
+  //register buttons to open folders
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Paks Folder', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn(path.join(GAME_PATH, UE5_ALT_PATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Binaries Folder', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn(path.join(GAME_PATH, BINARIES_PATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open UE4SS Mods Folder', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn( path.join(GAME_PATH, SCRIPTS_PATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open LogicMods Folder', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn(path.join(GAME_PATH, LOGICMODS_PATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Config Folder', async () => {
+    util.opn(CONFIG_PATH).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Saves Folder', async () => {
+    util.opn(SAVE_PATH).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Download UE4SS', () => {
+    downloadUe4ss(context.api, gameSpec).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open UE4SS Settings INI', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn(path.join(GAME_PATH, BINARIES_PATH, UE4SS_SETTINGS_FILEPATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open UE4SS mods.json', () => {
+    GAME_PATH = getDiscoveryPath(context.api);
+    util.opn(path.join(GAME_PATH, BINARIES_PATH, UE4SS_MODSJSON_FILEPATH)).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open PCGamingWiki Page', () => {
+    util.opn(PCGAMINGWIKI_URL).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'View Changelog', () => {
+    util.opn(path.join(__dirname, 'CHANGELOG.md')).catch(() => null);
+    }, () => {
+      const state = context.api.getState();
+      const gameId = selectors.activeGameId(state);
+      return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Submit Bug Report', () => {
+    util.opn(`${EXTENSION_URL}?tab=bugs`).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
+  context.registerAction('mod-icons', 300, 'open-ext', {}, 'Open Downloads Folder', () => {
+    util.opn(DOWNLOAD_FOLDER).catch(() => null);
+  }, () => {
+    const state = context.api.getState();
+    const gameId = selectors.activeGameId(state);
+    return gameId === GAME_ID;
+  });
 }
 
 //Main function
@@ -953,7 +1227,8 @@ function main(context) {
       callback: (loadOrder) => {
         if (previousLO === undefined) previousLO = loadOrder;
         if (loadOrder === previousLO) return;
-        context.api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
+        //context.api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
+        requestDeployment(context, spec);
         previousLO = loadOrder;
       },
       createInfoPanel: () =>
@@ -963,11 +1238,40 @@ function main(context) {
     });
   }
 
-  context.once(() => {
-    // put code here that should be run (once) when Vortex starts up
-
+  context.once(() => { // put code here that should be run (once) when Vortex starts up
+    context.api.onAsync('did-deploy', (profileId) => didDeploy(context.api, profileId)); //*/
   });
   return true;
+}
+
+const requestDeployment = (context, spec) => {
+  context.api.store.dispatch(actions.setDeploymentNecessary(spec.game.id, true));
+  context.api.sendNotification({
+    id: `${spec.game.id}-loadorderdeploy-notif`,
+    type: 'warning',
+    message: 'Deployment Required to Apply Load Order Changes',
+    allowSuppress: true,
+    actions: [
+      {
+        title: 'Deploy',
+        action: (dismiss) => {
+          deploy(context.api)
+          dismiss();
+        }
+      }
+    ],
+  });
+};
+
+async function didDeploy(api, profileId) { //run on mod deploy
+  const state = api.getState();
+  const profile = selectors.profileById(state, profileId);
+  const gameId = profile === null || profile === void 0 ? void 0 : profile.gameId;
+  if (gameId !== GAME_ID) {
+    return Promise.resolve();
+  }
+  api.dismissNotification(`${GAME_ID}-loadorderdeploy-notif`);
+  return Promise.resolve();
 }
 
 //export to Vortex